contrib: Revert to setting up the machine ID manually
Since we rebooted before first-boot-complete.target, the machine ID
generated during this boot may not be committed to the file system.
Thus this step must be performed manually.

This also makes sure the machine-id is generated before launching the
graphical environment, so the D-Bus and the rest of the system can live
happily.